requirements.txt
Django==1.7.1
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
django-celery-with-redis==3.0
django-tastypie==0.12.1
kombu==3.0.24
python-dateutil==2.3
python-mimeparse==0.1.4
pytz==2014.10
redis==2.10.3
six==1.8.0
wsgiref==0.1.2
